cells: Fix conb_1 and and2_0 NPC layers
The current version of the cells have "stubs" protruding from the center
toward the rails that are closer to the border than the min distance
for npc.

This means they can't be placed next to cells that have a middle npc zone
higher than theirs without causing a DRC issue.

To fix the issue, we increase the width of their middle zone to be as
wide as the widest other cell in the library to make sure this doesn't
happen.

The resulting cells were tested by placing them against (almost) every
other cells in the library in all possible orientation and running
klayout DRC deck to check npc. ( Excluded were double height cells and
some special cells like 'probe' )

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>